extract test from original source file
